Added specificRecipeView, updated bottomNav UI
Updated UI design to be more descriptive, and changed the icons to reflect the pages behind them.
Added a specific recipe view. For now, it just contains the title, image, instructions and description, but more is coming soon!
Added navigation for specific recipe view through both the bottomNav and by clicking on a recipe card